Bulk renaming to vc_stereo_ros2
