+++ include/vc_stereo_ros2/constants.h
@@ -6,11 +6,11 @@
 
 #include <Argus/Argus.h>
 
-namespace argus_stereo_sync {
+namespace vc_stereo_ros2 {
 
 static const Argus::Range<float> GAIN_RANGE(1, 44);
 static const Argus::Range<float> ISP_DIGITAL_GAIN_RANGE(1, 1);
 static const Argus::Range<uint64_t> EXPOSURE_TIME_RANGE(44000, 1000000);
 static const Argus::Size2D<uint32_t> STREAM_SIZE(1440, 1080);
 
-}  // namespace argus_stereo_sync
+}  // namespace vc_stereo_ros2
