+++ src/parameters.yaml
@@ -1,4 +1,4 @@
-argus_stereo_sync:
+vc_stereo_ros2:
 
   # Camera info params defaults to (empty) -- no camera info
   left_camera_info:
